Moved Invited Guest Model.
