+++ api/app/events/models.py
@@ -62,14 +62,3 @@ class EventRole(db.Model):
 
     def set_role(self, new_role):
         self.role = new_role
-
-
-class InvitedGuest(db.Model):
-    __tablename__ = "invited_guest"
-
-    id = db.Column(db.Integer(), primary_key=True)
-    event_id = db.Column(db.Integer(), db.ForeignKey(
-        "event.id"), nullable=False)
-    user_id = db.Column(db.Integer(), db.ForeignKey(
-        "app_user.id"), nullable=False)
-    role = db.Column(db.String(50), nullable=False)
